update get pools stats return
